pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shield-agent-mcp"
version = "0.1.0"
authors = [
  { name="gbvk", email="author@example.com" },
]
description = "A Hybrid-AI Security & Quality Sentinel for the MCP Era"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "google-generativeai>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
]

[project.scripts]
shield-agent = "shield_agent.cli:main"

[tool.setuptools]
packages = ["shield_agent"]
